test-deduplication in ColumnFamily & ListSnapshot
