+++ bigtable-hbase-2.x-parent/bigtable-hbase-2.x-integration-tests/src/test/java/com/google/cloud/bigtable/hbase/TestColumnFamilyAdmin.java
@@ -4,9 +4,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,20 +16,11 @@
 package com.google.cloud.bigtable.hbase;
 
 import org.apache.hadoop.hbase.HColumnDescriptor;
-import org.apache.hadoop.hbase.HTableDescriptor;
-import org.apache.hadoop.hbase.TableName;
-
-import java.io.IOException;
 
 /**
  * Tests creation and deletion of column families.
  */
 public class TestColumnFamilyAdmin extends AbstractTestColumnFamilyAdmin {
-
-  @Override
-  protected HTableDescriptor getTableDescriptor(TableName tableName) throws IOException {
-    return admin.getTableDescriptor(tableName);
-  }
 
   @Override
   protected void addColumn(byte[] columnName, int versions) throws Exception {
@@ -39,10 +30,5 @@ public class TestColumnFamilyAdmin extends AbstractTestColumnFamilyAdmin {
   @Override
   protected void modifyColumn(byte[] columnName, int versions) throws Exception {
     admin.modifyColumn(tableName, new HColumnDescriptor(columnName).setMaxVersions(versions));
-  }
-
-  @Override
-  protected void deleteColumn(byte[] columnName) throws Exception {
-    admin.deleteColumn(tableName, columnName);
   }
 }
